Replace 0.0.0.0 in addresses in references by the reference source address
This is needed to make direct join without using a masterserver working again where the real IPv4 address of the host is not known by the host.
This is just restoring the behavior from before the addition of IPv6 support.
IPv6 support changed it to add the address instead, but without ports and it didn't work.