C4Object::DoEnergy and C4Object::DoBreath: Parentesize C4MaxPhysical / 100
C4MaxPhysical % 100 is 0 anyway, so no precision is lost.
Instead, the minimal value of the other factor in the surrounding multiplication which will cause an overflow is increased by the factor of 100.
This should also fix a desync between (at least) GCC and MSVC, where MSVC will overflow with much smaller change values for DoEnergy/DoBreath than GCC.