Add "?" navigation modifier
This can be combined with navigation/access operators for safe navigation.
Can be used for ?., ?[i], ?-> and ?->~
In case the left hand side of ? is nil, the whole navigation/access chain is aborted and nil is returned.
The result of a ? modified navigation/access chain is always dereferenced.
This is due to the fact that the user of ? already knows that it may occasionally return non-ref nil and thus can't reliably use the result in reference context.