Remember the league login name permanently and the password in RAM
The password is cleared when the login fails or is cancelled.
To change the account without restarting the engine the password is also cleared when the player is removed in a lobby by the user themself.
This would break splitscreen in league, but currently only one player per host is allowed in league mode anyway.