C4StartupNetDlg: Fix update crash due to a race condition
C4UpdateDlg::CheckForUpdates enters a HandleMessage loop. If the update server
replies to the engine when the user has already exited from the network dialog,
it continues exection of the original C4StartupNetDlg::OnShown call, with this
being an invalid pointer.
This commit also only allows one update request at a time.