C4StartupNetDlg: Remove update button
C4UpdateDlg::CheckForUpdates runs the main loop, so startup GUI changes are
possible. This is guaranteed to break something as the network dialog can get
destroyed while the update check is still running; upon completion, it returns
into C4StartupNetDlg::OnShown, only with an invalid this pointer as the dialog
has already been destroyed. The automatic update check also caused the engine
to hand and crash in rather spectacular ways if the update server has problems.
Hence, this commit removes the update button from the network list; users
wanting to update are asked to use the manual update button in the about
screen as well as the automatic update check.