Leverage warnings for unhandled enum values in switches more often
This is achieved by not specifying a default case if all values should be handled already
and moving the code for the default case directly below the switch (only works correctly if all cases return (not break)).
Note: [[fallthrough]] at the end of a switch block (i.e. not in front of a case / default) issues a warning. That is why it is not used in some cases.