C4TexRef: Allocate texture in the constructor without uploading data
The common pattern is to create a surface and therefore textures, lock
it, load data into it and unlock it. This used to cause unnecessary
traffic to and from the GPU:

- an upload in Unlock() called from the constructor,
- a download in Lock(),
- another final upload in Unlock().

This commit changes the behavior so that the texture is always created
in the constructor (an exception is thrown if it fails) - and allocates
a texture on the GPU without uploading data to it. The instance is then
in a locked state, reducing the number of GPU data transfers to one,
which benefits performance.