Redo COM thread initialization and add C4Thread::Create
Previously the UI thread (aka the main thread) was initialized as a
multi-threaded apartment, which isn't what UI apartments are supposed to be
since they pump messages. This commit changes the main thread apartment to
a single-threaded apartment, ensures thread safety for ITaskbarList3 by only
accessing it via an agile reference from other threads and adds
C4Thread::Create to specify both the thread name and the thread apartment type.