Windows: Don't use an agile reference and switch back to a MTA for the main thread to fix a preloading mutex deadlock
If the preloading thread wanted to set the taskbar progress, it used an agile
reference to the task bar list object in the main thread to do so. However,
that required COM to call into the main thread, which required it to pump
messages, causing a deadlock if the main thread was waiting for the preloading
mutex that the preloading thread was holding.

This particular deadlock could have been fixed by not using an agile reference
without switching the main thread to a MTA, but a potential program deadlock
every time the main thread acquires a mutex is not acceptable.