C4Task: Remove PromiseTraits::IsNoexcept and refactor waiting for task cancellation on destruction
As cancellation throws an exception, IsNoexcept could cause terminate() to be called
just because the coroutine was already cancelled. Thus it is replaced by
TerminateOnException, which terminates for all exceptions with the exception
of CancelledException.

CancelAndWait() was also added, as well as task traits and a trait to make a task
wait for proper cancellation of its promise.